refactor(security/sandbox): flatten security/sandbox/__tests__ — rename 5 parallel suites (#2565 slice 9, final __tests__ slice)
Last collision slice of #2565. security/sandbox/__tests__ had 5 files
with basename collisions, each wrapping its sandbox-specific scope
in a single top-level describe block (parallel-suite style):

- command-allowlist.test.ts
- default-policies.test.ts
- env-sanitizer.test.ts
- sandbox-executor.test.ts
- sandbox-manager.test.ts

All 5 renamed with -extended.test.ts suffix.

Fix `from '../X'` static imports → `./X`. Fix `vi.mock('../X')` paths
(which my prior slices missed in the sed substitution) → `./X`. Fix
`await import('../X')` and `await import('../../X')` → local paths.
Remove the empty __tests__/ directory.

All 362 sandbox tests still pass.

This closes the basename-collision sweep portion of #2565. With this
PR, no `__tests__/` subdirectory remains under
packages/nexus-agents/src/. Each module's tests live as top-level
`<module>/<name>.test.ts` files with no two files in the same module
sharing a basename.
